Changed the interface and added class.
Device/Mountpoint:
The user will now not longer be determined for every method by a
parameter, but once by the constructor when the object is created.

FullLocation:
A new class representing a path on a host, possibly on a certain device
with a mountpoint, associated with a specific user. This class effectively
merges path, mountpoint and device into a single class.